+++ packages/api-rest/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.64](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-rest@4.0.63...@aws-amplify/api-rest@4.0.64) (2024-12-19)
+
+**Note:** Version bump only for package @aws-amplify/api-rest
+
 ## [4.0.63](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-rest@4.0.62...@aws-amplify/api-rest@4.0.63) (2024-12-12)
 
 **Note:** Version bump only for package @aws-amplify/api-rest